change float vars to double vars

+++ b3-dev-tu-flamant-benoit-armand-lelio/Arithmetic.cs
@@ -9,31 +9,31 @@ namespace b3_dev_tu_flamant_benoit_armand_lelio
     public class Arithmetic
     {
         //Arithmetic functions for a calculator
-        public static float Addition(float a, float b)
+        public static double Addition(double a, double b)
         {
             //Do an addition
             return a + b;
         }
 
-        public static float Substraction(float a, float b)
+        public static double Substraction(double a, double b)
         {
             //Do a substraction
             return a - b;
         }
 
-        public static float Multiplication(float a, float b)
+        public static double Multiplication(double a, double b)
         {
             //Do a multiplication
             return a * b;
         }
 
-        public static float Division(float a, float b)
+        public static double Division(double a, double b)
         {
             //Do a division
             return a / b;
         }
 
-        public static float Percentage(float a, float b)
+        public static double Percentage(double a, double b)
         {
             //b% of a
             return Multiplication(a, b) / 100;
